Fix cast warning

+++ api/samples/opcode_count.cpp
@@ -133,7 +133,7 @@ event_app_instruction(void *drcontext, void *tag, instrlist_t *bb, instr_t *inst
         static_cast<dr_spill_slot_t>(SPILL_SLOT_MAX + 1),
         IF_AARCHXX_(static_cast<dr_spill_slot_t>(SPILL_SLOT_MAX + 1)) &
             global_total_count,
-        bb_size,
+        (int)bb_size,
         /* TODO i#4215: DRX_COUNTER_LOCK is not yet supported on ARM. */
         IF_X86_ELSE(DRX_COUNTER_LOCK, 0));
